fog/requirements.txt
paho-mqtt==2.1.0
boto3>=1.35.0
